WIP: make handlers for files, combine the parsers, and start work on the actual loader
