+++ setup.py
@@ -9,5 +9,5 @@ setup(
     license='',
     author='Christopher J. Wright',
     author_email='wright1@email.sc.edu',
-    description=''
+    description='', requires=['tifffile']
 )
